#188 - test refactor
