+++ tests/extractors/videos.py
@@ -20,14 +20,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-import os
-
-from base import MockResponse
 from base import TestExtractionBase
-
-from goose.utils import FileHelper
-
-CURRENT_PATH = os.path.dirname(os.path.abspath(__file__))
 
 
 class ImageExtractionTests(TestExtractionBase):
